Rename remaining meeting time component
