--- bigbluebutton-html5/imports/ui/components/breakout-room/breakout-remaining-time/component.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-
-const BreakoutRemainingTime = props => (
-  <span data-test="timeRemaining">
-    {props.children}
-  </span>
-);
-
-
-export default BreakoutRemainingTime;
