+++ bigbluebutton-html5/imports/ui/components/breakout-room/meeting-remaining-time/component.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const MeetingRemainingTime = (props) => (
+  <span data-test="timeRemaining">
+    { props.children }
+  </span>
+);
+
+export default MeetingRemainingTime;
