+++ bigbluebutton-html5/imports/ui/components/notifications-bar/container.jsx
@@ -6,7 +6,7 @@ import _ from 'lodash';
 import Auth from '/imports/ui/services/auth';
 import { MeetingTimeRemaining } from '/imports/api/meetings';
 import Meetings from '/imports/api/meetings';
-import BreakoutRemainingTime from '/imports/ui/components/breakout-room/breakout-remaining-time/container';
+import MeetingRemainingTime from '../breakout-room/meeting-remaining-time/container';
 import Styled from './styles';
 import { layoutSelectInput, layoutDispatch } from '../layout/context';
 import { ACTIONS } from '../layout/enums';
@@ -176,7 +176,7 @@ export default injectIntl(withTracker(({ intl }) => {
 
     if (currentBreakout) {
       data.message = (
-        <BreakoutRemainingTime
+        <MeetingRemainingTime
           breakoutRoom={currentBreakout}
           messageDuration={intlMessages.breakoutTimeRemaining}
           timeEndedMessage={intlMessages.breakoutWillClose}
@@ -197,7 +197,7 @@ export default injectIntl(withTracker(({ intl }) => {
 
     if (underThirtyMin && !isBreakout) {
       data.message = (
-        <BreakoutRemainingTime
+        <MeetingRemainingTime
           breakoutRoom={meetingTimeRemaining}
           messageDuration={intlMessages.meetingTimeRemaining}
           timeEndedMessage={intlMessages.meetingWillClose}
