+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-content/breakout-room/component.jsx
@@ -4,7 +4,7 @@ import { defineMessages, injectIntl } from 'react-intl';
 import Icon from '/imports/ui/components/common/icon/component';
 import Styled from './styles';
 import { ACTIONS, PANELS } from '../../../layout/enums';
-import BreakoutRemainingTime from '/imports/ui/components/breakout-room/breakout-remaining-time/container';
+import MeetingRemainingTime from '/imports/ui/components/breakout-room/meeting-remaining-time/container';
 
 const intlMessages = defineMessages({
   breakoutTitle: {
@@ -66,7 +66,7 @@ const BreakoutRoomItem = ({
                   {intl.formatMessage(intlMessages.breakoutTitle)}
                 </Styled.BreakoutTitle>
                 <Styled.BreakoutDuration>
-                  <BreakoutRemainingTime
+                  <MeetingRemainingTime
                     messageDuration={intlMessages.breakoutTimeRemaining}
                     breakoutRoom={breakoutRoom}
                   />
